Add files via upload
Flappy Bird is a simple 2D arcade-style mobile game where the player controls a bird by tapping the screen to make it flap and fly. The objective is to navigate the bird through a series of green pipes without hitting them. The game is known for its minimalistic graphics, easy-to-learn but hard-to-master gameplay, and highly addictive nature. It was developed by Dong Nguyen and gained massive popularity due to its challenging mechanics and viral appeal.